% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{block_averaging}
\alias{block_averaging}
\title{Block averaging of non-overlapping submatrics in a given dataset}
\usage{
block_averaging(dat, block_size = 3, block_size_n, block_size_m)
}
\arguments{
\item{dat}{Dataset containing only numeric colu}

\item{block_size}{Default is 3.}

\item{block_size_n}{Vertical block size. If not specified, defaults to
block_size}

\item{block_size_m}{Horizontal block size. If not specified, defaults to
block_size}
}
\value{
Dataframe containing results of
}
\description{
Block averaging of non-overlapping submatrics in a given dataset
}
