% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{compare_rows_cols}
\alias{compare_rows_cols}
\title{Compare the dimensions of the original and new datasets}
\usage{
compare_rows_cols(dat_original, dat_comparison)
}
\arguments{
\item{dat_original}{Original dataset}

\item{dat_comparison}{Dataset to compare}
}
\value{
Indicator value: 0 = both dimensions equal, 1 = number of rows equal,
2 = number of columns equal, 3 = neither dimensions equal
}
\description{
Used to decide what comparison strategy to employ, as some
techniques are more sensitive to a change in a specific dimension than
others
}
