% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{get_pca_scores}
\alias{get_pca_scores}
\title{Perfrom PCR and}
\usage{
get_pca_scores(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
New dataset contain the scores matrix, ordered by the values of the
first principal component
}
\description{
Perfrom PCR and
}
