% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikegornik.R
\name{inellipse}
\alias{inellipse}
\title{Determine whether coordinates are enclosed by ellipse}
\usage{
inellipse(xdat, ydat, xellipse, yellipse)
}
\arguments{
\item{xdat}{numeric?}

\item{ydat}{numeric?}

\item{xellipse}{numeric?}

\item{yellipse}{numeric?}
}
\description{
Determine whether coordinates are enclose by ellipse
}
\examples{
\dontrun{
xellipse<-x
yellipse<-y
xdat<-speed[,a]
ydat<-speed[,b]
}
}

