% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{plot_qr}
\alias{plot_qr}
\title{Plot greyscale image of dataframe}
\usage{
plot_qr(dat)
}
\arguments{
\item{dat}{A dataframe scaled to range {0,...,255}, i.e. a dataframe that has
already had rescale_255() applied to it}
}
\value{
A "QR" style plot of the dataframe
}
\description{
Plot greyscale image of dataframe
}
