% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{rescale_255}
\alias{rescale_255}
\title{Convert to greyscale pixel values, i.e. range {0,...,255}}
\usage{
rescale_255(dat)
}
\arguments{
\item{dat}{Dataset containing only numeric variables}
}
\value{
Dataframe with all variables scaled to between
}
\description{
Convert to greyscale pixel values, i.e. range {0,...,255}
}
