% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikegornik.R
\name{spikegornik}
\alias{spikegornik}
\title{Despike ADV data using the GoringNikora phase-space procedure}
\usage{
spikegornik(fpath, outpath = NULL, maxcounter = 5, plotting = TRUE,
  plotwin = c(1, nrow(dt)))
}
\arguments{
\item{fpath}{file.path to ADV output where the column of interest is named "Speed"}

\item{outpath}{file.path to save results. optional.}

\item{maxcounter}{numeric number of smoothing passes}

\item{plotting}{logical output progress to plots?}

\item{plotwin}{numeric vector of length 2 specifying the plotting window}
}
\value{
data.frame datetime of class POSIXct and speed of class numeric
}
\description{
Despike ADV data
}
\details{
This function is nearly a direct translation of the MacVicar et al. (2014) SpikeGoringNikora.m script from MATLAB to R. Single value spikes are not removed if they are the same as a linear interpolation between adjacent points. Remove measurements where the SNR > 5 per the ADV manual?
}
\examples{
\dontrun{
fpath <- "inst/extdata/2014-10-14_R322_001_Z5-3.dat"
outpath <- "inst/extdata/2014-10-14_R322_001_Z5-3.csv"
res <- spikegornik(fpath)
res <- spikegornik(fpath, outpath)
plot(res, type = "l")
}
}
\author{
Joseph Stachelek
}
\references{
\url{https://github.com/macvicab/MITT}
}

