% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikegornik_cl.R
\name{spikegornik_cl}
\alias{spikegornik_cl}
\title{Command-line interface to the spikegornik function}
\usage{
spikegornik_cl(fpath, outpath)
}
\arguments{
\item{fpath}{file.path to ADV output where the column of interest is named "Speed"}

\item{outpath}{file.path to save results. optional.}
}
\description{
Command-line interface to the spikegornik function
}

