% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikegornik.R
\name{spikereplace}
\alias{spikereplace}
\title{Replace spikes}
\usage{
spikereplace(is_spike, sp)
}
\arguments{
\item{is_spike}{numeric vector of data length}

\item{sp}{numeric vector of raw data}
}
\description{
Replaces spikes
}

