% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikegornik.R
\name{spikegornik}
\alias{spikegornik}
\title{Despike ADV data using the GoringNikora phase-space procedure}
\usage{
spikegornik(fpath, outpath = NULL, plotting = TRUE, plotwin = c(1,
  nrow(dt)))
}
\arguments{
\item{fpath}{file.path to ADV output where the column of interest is named "Speed"}

\item{outpath}{file.path to save results. optional.}

\item{plotting}{logical output progress to plots?}

\item{plotwin}{numeric vector of length 2 specifying the plotting window}
}
\value{
data.frame datetime of class POSIXct and speed of class numeric
}
\description{
Despike ADV data
}
\details{
This function is nearly a direct translation of the MacVicar et al. (2014) SpikeGoringNikora.m script from MATLAB to R. Single value spikes are not removed if they are the same as a linear interpolation between adjacent points. Remove measurements where the SNR > 5 per the ADV manual?
}
\examples{
\dontrun{
#Newman test files
dt <- read.table("Z5_3_R3222001.dat", header = TRUE)
res <- spikegornik(dt, plotwin = c(nrow(dt)-200, nrow(dt)))
plot(dt[(nrow(dt)-200):nrow(dt), "SNR3"], type ="l")
plot(dt[(nrow(dt)-200):nrow(dt), "Speed"], type ="l")

par(mfrow = c(3, 1), mar = c(3, 3.9, 3, 1))
plot(dt[,"Speed"],type="l",main="Z5_3 Raw",ylab="Speed",xlab="")
plot(res[,"speed"], type = "l",ylim = c(0,max(dt[,"Speed"])),
main = "Z5_3 Despiked", ylab = "Speed", xlab = "")
plot(dt$SNR1, type = "l")
abline(h = 5, col = "red")

plot(dt[(nrow(dt)-200):nrow(dt),"Speed"],type="l",main="Z5_3 Raw",
ylab="Speed",xlab="")
plot(res[(nrow(res)-200):nrow(res),"speed"], type = "l", 
ylim = c(0,max(dt[,"Speed"])),main = "Z5_3 Despiked", ylab = "Speed", xlab = "")


dt <- read.table("Z5_1_R321001.dat", header = TRUE)
res <- spikegornik(dt, plotwin = c(nrow(dt)-200, nrow(dt)))
plot(dt[,"Speed"], type = "l", main = "Z5_1 Raw", ylab = "Speed",
xlab = "")
plot(res[,"speed"], type="l", ylim = c(0, max(dt[,"Speed"])),
main = "Z5_1 Despiked", ylab = "Speed", xlab = "")
}
}
\author{
Joseph Stachelek
}
\references{
\url{https://github.com/macvicab/MITT}
}

